Made EventController be responsible of dashboard, the search table and the get events as json hyperlink
