+++ src/main/resources/templates/dashboard.html
@@ -10,7 +10,7 @@
 
 <div class="header-container">
 <h1>Cauta dupa numele evenimentului: </h1>
-    <a id="json-events" href="http://localhost:8080//check" style="display: flex; justify-content: right;">Get events JSON</a>
+    <a id="json-events" href="http://localhost:8080/dashboard/events" style="display: flex; justify-content: right;">Get events JSON</a>
 <form action="keywords" method="get">
     <input name="keyword">
     <button>Cauta</button>
